--- a/aten/src/ATen/native/cuda/LegacyThrustHelpers.cu
+++ b/aten/src/ATen/native/cuda/LegacyThrustHelpers.cu
@@ -5,6 +5,8 @@
 #include <thrust/device_ptr.h>
 #include <thrust/execution_policy.h>
 #include <thrust/sort.h>
+#include <thrust/unique.h>
+#include <thrust/device_ptr.h>
 
 namespace at { namespace native {
 
@@ -29,5 +31,26 @@ void index_put_with_sort_kernel_thrust_helper(Tensor &linearIndex, Tensor &orig_
   auto sorted_data = device_ptr(sorted_indices.data_ptr<int64_t>());
   thrust::sort_by_key(policy, sorted_data, sorted_data + num_indices, orig_data, LTOp<int64_t>());
 }
+
+template<typename index_t>
+int embedding_renorm_cuda_unique_copy(Tensor &indices_contig, Tensor &unique_indices) {
+  cudaStream_t stream = at::cuda::getCurrentCUDAStream();
+  auto allocator = THCThrustAllocator(globalContext().lazyInitCUDA());
+  auto policy = thrust::cuda::par(allocator).on(stream);
+
+  using device_ptr = thrust::device_ptr<index_t>;
+
+  auto num_indices = indices_contig.numel();
+  auto indices_data = device_ptr(indices_contig.data_ptr<index_t>());
+  auto unique_data = device_ptr(unique_indices.data_ptr<index_t>());
+  auto end = thrust::unique_copy(policy, indices_data, indices_data + num_indices, unique_data);
+  auto num_unique_indices = static_cast<int>(end - unique_data);
+  return num_unique_indices;
+}
+
+template
+int embedding_renorm_cuda_unique_copy<int>(Tensor &indices_contig, Tensor &unique_indices);
+template
+int embedding_renorm_cuda_unique_copy<int64_t>(Tensor &indices_contig, Tensor &unique_indices);
 
 }}